# api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ResearchRequest": {
        "properties": {
          "k": {
            "default": 6,
            "description": "retrieval depth per sub-question",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "max_seconds": {
            "default": 60.0,
            "description": "wall-clock budget",
            "maximum": 300.0,
            "minimum": 5.0,
            "title": "Max Seconds",
            "type": "number"
          },
          "max_steps": {
            "default": 6,
            "description": "hard cap on retrieve+extract cycles",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "use_llm": {
            "default": true,
            "description": "use the LLM stages (falls back to heuristic keyless)",
            "title": "Use Llm",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "ResearchRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "CS/coding-specialized search backend for AI agents. Endpoints: `/search` (ranked evidence), `/v1/web_search` (drop-in web_search shape), `/research` (deep research), fetch/drill-down (`/source|/chunk|/claim/{id}`), `/graph`. Opaque handles: `chk_`/`doc_`/`clm_`/`ent_`. Errors use a structured envelope with a stable `code`. See `/contract` for the machine-readable descriptor and the MCP tool schemas.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/suppprith/moo/blob/main/LICENSE"
    },
    "summary": "CS/coding evidence search for AI agents",
    "title": "moo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chunk/{id}": {
      "get": {
        "description": "A chunk with full text, its document handle, and surrounding-chunk context.\nAccepts a `chk_<id>` handle (as emitted in agent-format sources) or bare rowid.",
        "operationId": "chunk_endpoint_chunk__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Chunk Endpoint Chunk  Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chunk Endpoint"
      }
    },
    "/claim/{id}": {
      "get": {
        "description": "A claim with confidence + its full evidence set (contradictions first) and\nextraction provenance. Accepts a `clm_<id>` handle or bare rowid.",
        "operationId": "claim_endpoint_claim__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Claim Endpoint Claim  Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Endpoint"
      }
    },
...
    "/graph": {
      "get": {
        "description": "Subgraph-for-query: seed entities from the query, their neighborhood, and\nthe claims attached to each node/edge. Sized for rendering (capped,\nexpandable via /graph/expand).",
        "operationId": "graph_graph_get",
        "parameters": [
          {
            "description": "natural-language query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "natural-language query",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 3,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cap",
            "required": false,
            "schema": {
              "default": 60,
              "maximum": 300,
              "minimum": 1,
              "title": "Cap",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Graph Graph Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Graph"
      }
    },
    "/graph/expand": {
      "get": {
        "description": "The next ring of edges for a node, for click-to-expand in the explorer.",
        "operationId": "graph_expand_graph_expand_get",
        "parameters": [
          {
            "description": "entity id or name",
            "in": "query",
            "name": "node",
            "required": true,
            "schema": {
              "description": "entity id or name",
              "title": "Node",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Graph Expand Graph Expand Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Graph Expand"
      }
    },
...
    "/research": {
      "post": {
        "description": "Run a full deep-research run (plan -> iterative loop -> cited report) and\nreturn the structured report. Budget is enforced end-to-end; a run that hits\na cap returns `status: partial`. Use POST /research/stream for progress, or\nGET /research/{id} to re-fetch.",
        "operationId": "research_post_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Post Research Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Post"
      }
    },
    "/research/stream": {
      "post": {
        "description": "Streaming (SSE) deep research: emits `plan`, a `progress` event per step\n(with why it was spawned), then a terminal `report` + `done` \u2014 or `error`.\nReuses the app.streaming event contract.",
        "operationId": "research_stream_research_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Stream"
      }
    },
    "/research/{run_id}": {
      "get": {
        "description": "Fetch a run's status + cited report for polling/resume. The report is\nreassembled deterministically (no LLM) from the persisted evidence.",
        "operationId": "research_get_research__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Get Research  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Get"
      }
    },
...
    "/source/{id}": {
      "get": {
        "description": "Full document behind a source: cleaned text + metadata + chunk handles.\nAccepts a `doc_<id>` handle or bare rowid.",
        "operationId": "source_endpoint_source__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Source Endpoint Source  Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Endpoint"
      }
    },
...
  }
}